Make test manifests useful and update README
This work updates the README to use the correct classes, and ensures
that the test cases represented in the README actually work.

Here we also remove the resource_permute function.  This should have
been removed long ago, since the module now depends on zleslie/permute.
If you don't have it, please go and get it.

https://github.com/xaque208/puppet-permute