BUGFIX: correct mfsbsd img options, prevent re-download of images
The regex /^mfsbsd(-se|-mini)$/ will not match on "pure" mfsbsd
images (those without -se and without -mini), which is not the intent
here.  /^mfsbsd(-se|-mini)?$/ will match correctly on the three options
applicable.

From the change of the login of the path variable to enable downloads
for mfsbsd > 10 in 6c4426f55d19d6893087fe8b4d41458f8fd19d28 follows the
necessity to adapt the logic of the "download-exec / creates =>"
statement to prevent multiple downloads.

	modified:   manifests/images.pp
	modified:   manifests/images/mfsbsd.pp