Fix #35 Begin unit testing
This should have been long ago.  Currently there are no unittests to
ensure version compatibility, or that the module functions in the
correct way.  Here we begin unit testing, and make a couple other
changes to the testing structure to more closely match the other modules
under my maintenance.  This should allow for easier maintainability, and
greater assurance that the module is still working on newer versions of
puppet, etc.