Update the readme for pxe::installer usage
Without this work, users are still led to believe the permute is the
preferred way to create PXE installers with this module.  The permute
function was written in a time before iteration.  As such, the new
language constructs have replaced the need for any to ever use the
permute function again.

Here we update our documentation to provide a modern example of how
resource could be generated without the use of the permute function.